split traits module into layers and sources modules
