+++ src/style/layers/mod.rs
@@ -3,9 +3,11 @@ mod fill;
 mod id;
 mod line;
 mod symbol;
+mod traits;
 
 pub use circle::CircleLayer;
 pub use fill::FillLayer;
 pub use id::LayerId;
 pub use line::{LineCap, LineJoin, LineLayer};
 pub use symbol::{SymbolAnchor, SymbolLayer};
+pub use traits::Layer;
