+++ src/style/layers/traits.rs
@@ -1,7 +1,7 @@
-use crate::bridge::{ffi, layers, sources};
-use crate::style::{CircleLayer, FillLayer, GeoJsonSource, LineLayer, SymbolLayer};
+use crate::bridge::{ffi, layers};
+use crate::style::{CircleLayer, FillLayer, LineLayer, SymbolLayer};
 
-use sealed::{IntoLayer, IntoSource};
+use sealed::IntoLayer;
 
 mod sealed {
     use crate::bridge::ffi;
@@ -10,30 +10,7 @@ mod sealed {
         fn layer_id(&self) -> &str;
         fn into_layer(self) -> cxx::UniquePtr<ffi::CxxLayer>;
     }
-
-    pub trait IntoSource {
-        fn source_id(&self) -> &str;
-        fn into_source(self) -> cxx::UniquePtr<ffi::CxxSource>;
-    }
 }
-
-/// A style source type that can be added to a [`Style`](crate::Style).
-///
-/// This trait is sealed; only source types provided by this crate can implement
-/// it.
-pub trait Source: IntoSource {}
-
-impl IntoSource for GeoJsonSource {
-    fn source_id(&self) -> &str {
-        self.source_id()
-    }
-
-    fn into_source(self) -> cxx::UniquePtr<ffi::CxxSource> {
-        sources::geojson_into_source(self.into_inner())
-    }
-}
-
-impl Source for GeoJsonSource {}
 
 /// A style layer type that can be added to a [`Style`](crate::Style).
 ///
