+++ src/style/sources/mod.rs
@@ -1,5 +1,7 @@
 mod geojson_source;
 mod id;
+mod traits;
 
 pub use geojson_source::GeoJsonSource;
 pub use id::SourceId;
+pub use traits::Source;
